+++ template-sample-app/template/src/components/common/Header.js
@@ -93,7 +93,7 @@ function Desktop() {
       </Link>
       <Divider orientation='vertical' className={classes.divider} light></Divider>
       <NavigationMenu pathname={pathname} />
-      <Grid container item xs justify='flex-end'>
+      <Grid container item xs justifyContent='flex-end'>
         <UserMenu />
       </Grid>
     </Hidden>
@@ -163,7 +163,7 @@ function Mobile() {
         }}
       >
         <Toolbar variant='dense' />
-        <Grid container direction='column' justify='space-between' item xs>
+        <Grid container direction='column' justifyContent='space-between' item xs>
           <NavigationMenu column={true} />
         </Grid>
       </Drawer>