+++ template-skeleton-cn/template/src/components/common/DefaultView.js
@@ -1,0 +1,10 @@
+import Header from 'components/common/Header';
+
+export default function DefaultView({ children }) {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  );
+}