+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Refactor to simplify routes file in all templates [#263](https://github.com/CartoDB/carto-react-template/pull/263)
 - Update deck.gl version to 8.5.0 [#264](https://github.com/CartoDB/carto-react-template/pull/264)
 - Fix Material-UI warnings on justify property and createTheme import [#264](https://github.com/CartoDB/carto-react-template/pull/264)